Made some much needed changes to the respawning datapack
Removed FreezePlayers datapack and removed all references to it.

Now instead repeatedly teleports spectating players to "0 200 0". Effectively freezing them in place.

It also now sets the spectatorsGenerateChunks gamerule to "true" on load.

The spawnRadius gamerule now gets set on every load. This is to prevent it from resetting down the line.

When respawning the player now gets teleported to a random location (within 5000 blocks)