Bundled 'GUI Reimagined' and disabled it in the metadata
This is so that I can rename the file to exclude the "§" symbol. Without this workaround it is not possible to change the default load order of this resourcepack using 'Resource Pack Overrides'.

Also removed it as a global pack. Now it should load just under the Insomnia resources packs.

Also made it so that you cannot change the order between 'FreshAnimations' and 'Tissou's Zombie Pack'.