Fixing server side issues
'Extra Origins' crashes the server when using Fabric-0.15.x. To fix this i have decided to use Fabric-0.14.25 on just the server in order to avoid this issue.

Also removed 'DimThread' due to it spamming errors related to 'Mining Dimensions'.

Also replaced 'Better End Cities Better End' mod on the server side with its datapack counterpart. This is because the mod version needs Fabric-0.15.3 or higher.